Re-arranging product metadata readers.
